requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
opencv-python>=4.8.0
pillow>=9.5.0
numpy>=1.24.0
gradio>=3.40.0
huggingface-hub>=0.16.0
safetensors>=0.3.0
scipy>=1.10.0
# Training dependencies
tensorboard>=2.13.0
tqdm>=4.65.0
yt-dlp>=2023.0.0
einops>=0.6.1